+++ android/src/main/kotlin/com/squareup/square_mobile_payments_sdk/modules/SettingsModule.kt
@@ -61,9 +61,9 @@ class SettingsModule {
         @JvmStatic
         fun getTrackingConsentState(result: MethodChannel.Result) {
             val consentState = settingsManager.trackingConsentState
-            result.success(consentState.name) // Or map it to a specific string/integer if needed
+            result.success(consentState.name)
         }
-
+        @JvmStatic
         @JvmStatic
         fun updateTrackingConsent(result: MethodChannel.Result, granted: Boolean) {
             settingsManager.updateTrackingConsent(granted) // Assuming the method name is similar